# backend/docs/openapi.json
{"openapi":"3.0.3","info":{"title":"BookPath API","version":"1.0.0","description":"API documentation for BookPath application (v1)","license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"}},"servers":[{"url":"http://localhost:3001/api/v1","description":"Development server"},{"url":"https://api.bookpath.eu/api/v1","description":"Production server"}],"tags":[{"name":"Auth","description":"Authentication endpoints"},{"name":"Users","description":"User management endpoints"},{"name":"Books","description":"Book discovery and metadata"},{"name":"Collections","description":"Book collection management"},{"name":"Billing","description":"Subscription and billing"},{"name":"Metrics","description":"Operational metrics (admin)"},{"name":"SEO","description":"SEO utility endpoints"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"ErrorResponse":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Error message"}}},"Pagination":{"type":"object","properties":{"currentPage":{"type":"integer","example":1},"totalPages":{"type":"integer","example":10},"totalResults":{"type":"integer","example":200},"hasNextPage":{"type":"boolean","example":true},"hasPreviousPage":{"type":"boolean","example":false}}},"Book":{"type":"object","properties":{"id":{"type":"string","description":"Internal or external ID"},"title":{"type":"string"},"authors":{"type":"array","items":{"type":"string"}},"description":{"type":"string"},"coverImage":{"type":"string","format":"uri"},"firstPublishYear":{"type":"integer"},"subjects":{"type":"array","items":{"type":"string"}},"category":{"type":"string","description":"Normalized category (e.g., Sci-Fi, Fantasy)"},"condition":{"type":"string","enum":["new","used","unknown"]},"amazonLink":{"type":"string","nullable":true}}},"BookSearchResponse":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"type":"array","items":{"$ref":"#/components/schemas/Book"}},"pagination":{"$ref":"#/components/schemas/Pagination"},"sources":{"type":"object","properties":{"openLibrary":{"type":"boolean"},"gutendex":{"type":"boolean"},"googleBooks":{"type":"boolean"},"booklooker":{"type":"boolean"}}},"errors":{"type":"array","items":{"type":"string"},"nullable":true}}},"User":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"email":{"type":"string","format":"email"},"role":{"type":"string"}}},"BookCollection":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"isPublic":{"type":"boolean"},"books":{"type":"array","items":{"$ref":"#/components/schemas/Book"}}}}}},"paths":{"/auth/register":{"post":{"operationId":"authRegister","tags":["Auth"],"summary":"Register a new user","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["username","email","password"],"properties":{"username":{"type":"string"},"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password"}}}}}},"responses":{"201":{"description":"Created"},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict"}}}},"/auth/login":{"post":{"operationId":"authLogin","tags":["Auth"],"summary":"Login user","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password"}}}}}},"responses":{"200":{"description":"OK"},"401":{"description":"Unauthorized"}}}},"/auth/logout":{"post":{"operationId":"authLogout","tags":["Auth"],"summary":"Logout user","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"OK"},"401":{"description":"Unauthorized"}}}},"/auth/refresh-token":{"post":{"operationId":"authRefreshToken","tags":["Auth"],"summary":"Refresh JWT token","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["refreshToken"],"properties":{"refreshToken":{"type":"string"}}}}}},"responses":{"200":{"description":"OK"},"401":{"description":"Unauthorized"}}}},"/books/search":{"get":{"operationId":"booksSearch","tags":["Books"],"summary":"Search for books across multiple sources","parameters":[{"in":"query","name":"title","schema":{"type":"string"}},{"in":"query","name":"author","schema":{"type":"string"}},{"in":"query","name":"page","schema":{"type":"integer","default":1,"minimum":1}},{"in":"query","name":"category","schema":{"type":"string"}},{"in":"query","name":"condition","schema":{"type":"string","enum":["new","used"]}},{"in":"query","name":"sort","schema":{"type":"string","enum":["newest","author_az"]}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookSearchResponse"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/books/search/advanced":{"get":{"operationId":"booksAdvancedSearch","tags":["Books"],"summary":"Advanced book search","security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"title","schema":{"type":"string"}},{"in":"query","name":"author","schema":{"type":"string"}},{"in":"query","name":"genre","schema":{"type":"string"}},{"in":"query","name":"page","schema":{"type":"integer","default":1,"minimum":1}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookSearchResponse"}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"402":{"description":"Requires subscription"}}}},"/books/{id}":{"get":{"operationId":"booksGetById","tags":["Books"],"summary":"Get book by ID","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/Book"}}}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/books/author/{authorId}":{"get":{"operationId":"booksGetAuthor","tags":["Books"],"summary":"Get author details","parameters":[{"in":"path","name":"authorId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/collections":{"get":{"operationId":"collectionsList","tags":["Collections"],"summary":"Get user's collections","security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"page","schema":{"type":"integer","default":1}},{"in":"query","name":"limit","schema":{"type":"integer","default":10}}],"responses":{"200":{"description":"OK"}}},"post":{"operationId":"collectionsCreate","tags":["Collections"],"summary":"Create a new collection","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["name"],"properties":{"name":{"type":"string"},"description":{"type":"string"},"isPublic":{"type":"boolean"}}}}}},"responses":{"201":{"description":"Created"}}}},"/collections/{id}":{"get":{"operationId":"collectionsGetById","tags":["Collections"],"summary":"Get collection by ID","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"put":{"operationId":"collectionsUpdate","tags":["Collections"],"summary":"Update collection","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"isPublic":{"type":"boolean"},"category":{"type":"string"},"color":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}}}}}}},"responses":{"200":{"description":"OK"}}},"delete":{"operationId":"collectionsDelete","tags":["Collections"],"summary":"Delete collection","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/collections/by-category/{category}":{"get":{"operationId":"collectionsGetByCategory","tags":["Collections"],"summary":"Get collections by category","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"category","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/collections/{id}/share":{"post":{"operationId":"collectionsShare","tags":["Collections"],"summary":"Generate shareable link","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/collections/{id}/stats":{"get":{"operationId":"collectionsStats","tags":["Collections"],"summary":"Get collection statistics","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/collections/{id}/search":{"get":{"operationId":"collectionsSearchInCollection","tags":["Collections"],"summary":"Search books in a collection","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}},{"in":"query","name":"query","schema":{"type":"string"}},{"in":"query","name":"status","schema":{"type":"string"}},{"in":"query","name":"genre","schema":{"type":"string"}},{"in":"query","name":"favorite","schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/users/data-export":{"get":{"operationId":"usersDataExport","tags":["Users"],"summary":"Export user data (GDPR compliance)","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"User data export","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/User"},"collections":{"type":"array","items":{"$ref":"#/components/schemas/BookCollection"}},"exportedAt":{"type":"string","format":"date-time"}}}}}}}},"401":{"description":"Unauthorized"}}}},"/users/delete-account":{"delete":{"operationId":"usersDeleteAccount","tags":["Users"],"summary":"Delete user account and all associated data (GDPR compliance)","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["password"],"properties":{"password":{"type":"string","description":"Current password for confirmation"}}}}}},"responses":{"200":{"description":"Account deleted successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"}}}}}},"400":{"description":"Invalid password"},"401":{"description":"Unauthorized"}}}},"/users/notification-preferences":{"get":{"operationId":"usersGetNotificationPreferences","tags":["Users"],"summary":"Get user notification preferences","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Notification preferences","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"emailNotifications":{"type":"object","properties":{"marketing":{"type":"boolean"},"productUpdates":{"type":"boolean"},"securityAlerts":{"type":"boolean"},"collectionShares":{"type":"boolean"}}}}}}}}}},"401":{"description":"Unauthorized"}}},"put":{"operationId":"usersUpdateNotificationPreferences","tags":["Users"],"summary":"Update user notification preferences","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"emailNotifications":{"type":"object","properties":{"marketing":{"type":"boolean"},"productUpdates":{"type":"boolean"},"securityAlerts":{"type":"boolean"},"collectionShares":{"type":"boolean"}}}}}}}},"responses":{"200":{"description":"Preferences updated","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"}}}}}},"400":{"description":"Invalid preferences"},"401":{"description":"Unauthorized"}}}},"/billing/checkout":{"post":{"operationId":"billingCheckout","tags":["Billing"],"summary":"Create Stripe Checkout session","security":[{"bearerAuth":[]}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"type":"object","properties":{"priceId":{"type":"string","description":"Optional Stripe Price ID (defaults to env)"}}}}}},"responses":{"200":{"description":"Created","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"url":{"type":"string","format":"uri"},"id":{"type":"string"}}}}}},"401":{"description":"Unauthorized"}}}},"/billing/portal":{"post":{"operationId":"billingPortal","tags":["Billing"],"summary":"Create Stripe Billing Portal session","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Created","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"url":{"type":"string","format":"uri"}}}}}},"401":{"description":"Unauthorized"}}}},"/billing/webhook":{"post":{"operationId":"billingWebhook","tags":["Billing"],"summary":"Stripe webhook endpoint","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","description":"Stripe webhook event payload"}}}},"parameters":[{"in":"header","name":"stripe-signature","required":true,"schema":{"type":"string"},"description":"Stripe webhook signature"}],"responses":{"200":{"description":"Webhook processed successfully"},"400":{"description":"Invalid signature or payload"}}}},"/seo/collections/{shareableLink}":{"get":{"operationId":"seoCollectionPage","tags":["SEO"],"summary":"SEO HTML for public collection","parameters":[{"in":"path","name":"shareableLink","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"HTML","content":{"text/html":{"schema":{"type":"string"}}}},"404":{"description":"Not found"}}}},"/metrics":{"get":{"operationId":"metrics","tags":["Metrics"],"summary":"Prometheus metrics","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Metrics text","content":{"text/plain":{"schema":{"type":"string"}}}}}}}},"security":[{"bearerAuth":[]}]}